requirements.txt
flask==2.0.1
python-dotenv==0.19.0
google-generativeai==0.3.0
flask-sqlalchemy==3.0.2
flask-migrate==4.0.5
